Major refactor - Moving Components to new Directory
