+++ anubis-family/src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {DailyTasksComponent} from './component/daily-tasks/daily-tasks.component';
-import {WeeklyTasksComponent} from './weekly-tasks/weekly-tasks.component';
+import {WeeklyTasksComponent} from './component/weekly-tasks/weekly-tasks.component';
 import {DailyTaskComponent} from './component/daily-tasks/daily-task/daily-task.component';
-import {RegisterComponent} from './register/register.component';
-import {CalendarComponent} from './calendar/calendar.component';
-import {ProfileComponent} from './profile/profile.component';
+import {RegisterComponent} from './component/register/register.component';
+import {CalendarComponent} from './component/calendar/calendar.component';
+import {ProfileComponent} from './component/profile/profile.component';
 
 const routes: Routes = [
   {path: 'register', component: RegisterComponent},
