+++ anubis-family/src/app/app.module.ts
@@ -4,24 +4,24 @@ import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 
 import {AppMaterialModule} from './app-material.module';
 import {AppRoutingModule} from './app-routing.module';
-import {AppComponent} from './app.component';
+import {AppComponent} from './component/app.component';
 
 import {LayoutModule} from '@angular/cdk/layout';
 
-import {NavigationComponent} from './navigation/navigation.component';
+import {NavigationComponent} from './component/navigation/navigation.component';
 import {DailyTasksComponent} from './component/daily-tasks/daily-tasks.component';
-import {WeeklyTasksComponent} from './weekly-tasks/weekly-tasks.component';
+import {WeeklyTasksComponent} from './component/weekly-tasks/weekly-tasks.component';
 import {HttpClientModule} from '@angular/common/http';
 import {DailyTaskComponent} from './component/daily-tasks/daily-task/daily-task.component';
-import {RegisterComponent} from './register/register.component';
+import {RegisterComponent} from './component/register/register.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {RegisterDialogComponent} from './register/register-dialog.component';
+import {RegisterDialogComponent} from './component/register/register-dialog.component';
 import {authInterceptorProviders} from './interceptor/auth.interceptor';
-import {LoginComponent} from './login/login.component';
-import {LoginDialogComponent} from './login/login-dialog/login-dialog.component';
-import {CalendarComponent} from './calendar/calendar.component';
-import {ProfileComponent} from './profile/profile.component';
-import {ProfileEditComponent} from './profile/profile-edit/profile-edit.component';
+import {LoginComponent} from './component/login/login.component';
+import {LoginDialogComponent} from './component/login/login-dialog/login-dialog.component';
+import {CalendarComponent} from './component/calendar/calendar.component';
+import {ProfileComponent} from './component/profile/profile.component';
+import {ProfileEditComponent} from './component/profile/profile-edit/profile-edit.component';
 import {DailyTaskCreateComponent} from './component/daily-tasks/daily-task-create/daily-task-create.component';
 import {MatSelectModule} from '@angular/material/select';
