+++ anubis-family/src/app/component/login/login.component.ts
@@ -1,8 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {AbstractControl, FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {AuthService} from '../service/auth.service';
-import {TokenStorageService} from '../service/token-storage.service';
-import {ProfileService} from '../service/profile.service';
+import {AuthService} from '../../service/auth.service';
+import {TokenStorageService} from '../../service/token-storage.service';
+import {ProfileService} from '../../service/profile.service';
 
 @Component({
   selector: 'app-login',
