+++ anubis-family/src/app/component/register/register.component.less
