+++ anubis-family/src/test/app.component.spec.ts
@@ -1,6 +1,6 @@
 import {TestBed} from '@angular/core/testing';
 import {RouterTestingModule} from '@angular/router/testing';
-import {AppComponent} from '../app/app.component';
+import {AppComponent} from '../app/component/app.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
